One public header per library; tests run as build steps
Four changes that all pull the same direction: fewer things a reader has to
know, and fewer commands they have to remember.

**One public header per library.** A library's API is
`src/<lib>/include/scav/scav_<lib>.h`, and that directory is its only PUBLIC
include path -- everything else under `src/<lib>/` needs `-Isrc`, which is
PRIVATE to the library and its own tests. A consumer cannot name a private
header, in the build tree or an installed one.

Core's API was briefly split eight ways, which is what PRD 3.3 already said not
to do. It is one header now: scav_core.h, sectioned, with every function
carrying the prefix of its section (parse_, lex_, source_text_, diag_, syntax_,
string_) so the section list is a table of contents and a symbol is
self-locating. Three public headers exist in the whole project, and a core user
includes one. func.install_consumer pins that by name, links scav::core as well
as scav::toy, compiles every public header against nothing but the install
prefix, and parses a chart through them -- which is what catches a public header
that quietly includes a private one.

Names that only made sense from the inside are gone. `normalize.h` read like a
vector norm; it is the source-text section, covering BOM, line endings, UTF-8
and NFC on read. Interning moved out of the public surface entirely: a client
reads a finalized StringPool, and only the parser builds one.

**Tests are build steps.** Every test, unit and functional, is an
add_custom_command whose output is a stamp, wired into ALL. Building is testing,
a green build cannot hide a red test, and a second build back to back is
`ninja: no work to do` in zero seconds. That last property is why CTest is gone
rather than wrapped: ctest has no notion of a test being up to date, so it
re-runs everything every invocation and "build, then test" can never be
incremental. Each functional test declares the files it reads, so editing
ci.yml re-runs only the test that reads it. A failing test deletes its stamp and
fails the build, so the next build retries rather than trusting a stale pass.
Two things needed real handling: the deliberately-failing case inverts, so it
runs through a cmake -P script that succeeds only on a non-zero exit; and
ctest's noTestsAction=error became a configure-time check that fires earlier.

**size_t at the boundary.** Public buffer lengths take size_t, because that is
what the type is for and what .size() returns. They are range-checked into
uint32_t at each entry with a DocumentTooLarge diagnostic, because Span and
StrRef are {uint32 off, len} at the ABI (PRD 16) and PRD 6 bans size_t from
value computation. Auditing that turned up the real bug: 60 unchecked narrowing
sites, against a PRD 6 rule naming one `narrow<T>()` helper checked in every
build. The helper did not exist. It does now, every site goes through it or
through narrow_clamp, and both are tested along with the new diagnostic.

**Library-local tooling lives with its library.** gen_unicode_tables.py is
core/lang's alone, so it sits beside the code it generates and writes beside
itself, driven by a local scav_core_unicode_tables target. Deliberately not in
ALL: the .inc files are committed so an ordinary build needs neither network nor
Python, and the Unicode version is a determinism input rather than something a
build may move underneath you. tools/ is project-wide tooling only.

Also removes a stray CTest scratch directory that reached the tree in the
previous commit, and gitignores it.

Verified with one compiler across the whole sweep: debug, release, testable,
ASan, UBSan, TSan, coverage, clang-tidy and clang-format all green; 195 cases,
146,089 assertions; 100% of lines on all 9 production files; second build a
0s no-op.

+++ src/core/model/lookup_map.h
@@ -14,7 +14,6 @@
 // or gets hashed (PRD 4.1's test).
 
 #include <cstddef>
-#include <cstdint>
 #include <functional>
 #include <string>
 #include <string_view>
@@ -68,10 +67,10 @@ class LookupMap {
     return map.emplace(std::move(key), std::move(value)).second;
   }
 
-  [[nodiscard]] uint32_t size() const { return static_cast<uint32_t>(map.size()); }
+  [[nodiscard]] size_t size() const { return map.size(); }
   [[nodiscard]] bool empty() const { return map.empty(); }
   void clear() { map.clear(); }
-  void reserve(uint32_t n) { map.reserve(n); }
+  void reserve(size_t n) { map.reserve(n); }
 
  private:
   std::unordered_map<K, V, Hash, Eq> map;
